fix(serve): validate cordova-build options to provide defaults
Using `validateBuilderOptions()` is key--it fills in default values from
the `schema.json` file. In this case, the new `cordovaAssets` option in
the cordova-build Builder defaults to `true`, but if cordova-serve uses
it without validating the options, the default is lost, resulting in
"cordova not available" issues.

fixes https://github.com/ionic-team/ionic-cli/issues/3720